requirements.txt
altgraph==0.17.3
automata-lib==6.0.2
autopep8==1.7.0
Babel==2.11.0
certifi==2022.9.24
charset-normalizer==2.1.1
future==0.18.3
graphviz==0.20.1
greenlet==2.0.2
idna==3.4
jedi==0.18.1
keyboard==0.13.5
msgpack==1.0.4
networkx==2.8.7
parso==0.8.3
pefile==2022.5.30
Pillow==9.5.0
pluggy==1.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycodestyle==2.9.1
pydot==1.4.2
pydot-ng==2.0.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pynput==1.7.6
pynvim==0.4.3
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
pytz==2022.6
pywin32-ctypes==0.2.0
requests==2.28.1
six==1.16.0
SQLAlchemy==2.0.16
tkcalendar==1.6.1
toml==0.10.2
ttkbootstrap==1.10.1
typing_extensions==4.6.3
ujson==5.5.0
urllib3==1.26.12
